+++ pkg/sdk/go/domains_test.go
@@ -14,9 +14,9 @@ import (
 	httpapi "github.com/absmach/magistrala/auth/api/http/domains"
 	authmocks "github.com/absmach/magistrala/auth/mocks"
 	internalapi "github.com/absmach/magistrala/internal/api"
-	"github.com/absmach/magistrala/internal/apiutil"
 	"github.com/absmach/magistrala/internal/testsutil"
 	mglog "github.com/absmach/magistrala/logger"
+	"github.com/absmach/magistrala/pkg/apiutil"
 	mgclients "github.com/absmach/magistrala/pkg/clients"
 	"github.com/absmach/magistrala/pkg/errors"
 	svcerr "github.com/absmach/magistrala/pkg/errors/service"
